Restructuring TierList components
