--- src/components/ChampionEntryModal/index.tsx
@@ -1,3 +1,0 @@
-import ChampionEntryModal from './ChampionentryModal';
-
-export default ChampionEntryModal;
