+++ src/components/TierList/ChampionList/ChampionListEntryModal/index.tsx
@@ -1,0 +1,3 @@
+import ChampionListEntryModal from './ChampionListEntryModal';
+
+export default ChampionListEntryModal;
