+++ src/components/TierList/TierListModal/TierListModal.tsx
