+++ src/components/TierList/TierListModal/index.tsx
